fairlens_backend-main/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.0
